updated logout
